Remove interface_exists() and class_exists() calls
The calls to interface_exists() and class_exists() are unnecessary. They just force calling the autoloader, which can be expensive. In order to use instanceof, it is not necessary to check if the class exists. If the class does not exist in memory, the $argument can never be a sub class of said class.