PDOStatement mocking was not working
When mocking a PDOStatement an exception was thrown:
"PDOException: You cannot serialize or unserialize PDO instances"
This used to work in alpha2 and was broken in alpha3. This is a naive
patch that just catches the exception and uses the old way of creating
the object. I could not duplicate this happening in general with an
exception being thrown in __wakeup() or unserialize(), if implements
Serializable.