Add phpDoc to loaders and generated mock classes
This enables checking of mocked return values via Augmented Types:
https://github.com/box/augmented_types

Augmented types allow for an extra level of automated checking of programmer
errors by ensuring that values conform to the asserted types in the phpDoc.
This will allow those same errors to be caught when they are made on mock objects
thereby indicating an invalid test.