Adding support for scalar type hints and return types.
Return types are particularly tricky as now PHP does not allow returning null for hinted return types. This results in a behavioral change for Phake. Now, by default, the default value for each given return type will be the default answer. The defaults are achieved by essentially the "empty" value for each built-in type: 0, 0.0, '', false, array(). A hoop function is the default value for a callable, and if the return type is another class, a mock of that class will be returned.