Private methods are not mockable anymore
Private methods were mockable because of a bug in
`Phake\ClassGenerator\MockClass` where a NOT bitwise operations (~)
was used where the method do not support it. Instead of not retrieving
the final methods (~ReflectionMethod::IS_FINAL) all privates methods
where included.

There is no point in mocking a private method anyway since the only way
to call it is via reflection.